add a condition to check if the current user match with the author of the food post to delete

+++ controllers/food.js
@@ -89,6 +89,7 @@ exports.editFood = (req, res, next) => {
             if(food.userId.toString() !== req.user.id){
                 return res.status(403).json({error: 'You are not allow to edit this food post'});
             }
+            
             food.name = name;
             food.price = price;
             food.location = location;
@@ -112,6 +113,11 @@ exports.removeFood = (req, res, next) => {
             if(!food){
                 return res.status(404).json({error: 'Food post not found'});
             }
+            
+            if(food.userId.toString() !== req.user.id){
+                return res.status(403).json({error: 'You are not allow to delete this food post'});
+            }
+            
             return Food.findByIdAndRemove(foodId);
         })
         .then(result => {
